+++ src/query/query.ts
@@ -1,4 +1,4 @@
-import { Doc } from "../util/doc-types.ts";
+import { DocBase, FormatName } from "../util/doc-types.ts";
 import { DEFAULT_QUERY, Query, QueryFilter } from "./query-types.ts";
 
 import { deepEqual } from "../util/misc.ts";
@@ -196,7 +196,10 @@ export function cleanUpQuery(inputQuery: Query): CleanUpQueryResult {
   };
 }
 
-export function docMatchesFilter(doc: Doc, filter: QueryFilter): boolean {
+export function docMatchesFilter<
+  FormatType extends FormatName,
+  DocType extends DocBase<FormatType>,
+>(doc: DocType, filter: QueryFilter): boolean {
   // Does the doc match the filters?
   if (filter.path !== undefined && doc.path !== filter.path) return false;
   if (
@@ -229,6 +232,9 @@ export function docMatchesFilter(doc: Doc, filter: QueryFilter): boolean {
   ) {
     return false;
   }
+
+  // TODOM3: How do we check content length, now that not every doc has a content field?
+  /*
   const contentLength = stringLengthInBytes(doc.content);
   if (
     filter.contentLength !== undefined &&
@@ -248,11 +254,15 @@ export function docMatchesFilter(doc: Doc, filter: QueryFilter): boolean {
   ) {
     return false;
   }
+  */
   return true;
 }
 
 /** Return whether a document is expired or not */
-export function docIsExpired(doc: Doc, now?: number) {
+export function docIsExpired<
+  FormatType extends FormatName,
+  DocType extends DocBase<FormatType>,
+>(doc: DocType, now?: number) {
   const nowToUse = now || Date.now() * 1000;
 
   if (doc.deleteAfter === null) {