+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next
+
+- BREAKING: Replica's constructor arguments changed. It now takes a single
+  options argument instead of three. In these new options, the share address
+  does not need to be given, and it now accepts an array of format validators
+  rather than a single one.
+
 ## v9.3.2
 
 - Updated earthstar_streaming_rcp to v5.0.1